Refactor sdpa fw exp function to shared file
